requirements.txt
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.7
python-binance==0.7.11
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
zope.interface==5.4.0
